requirements.txt
asgiref==3.4.1
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.9
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==4.0
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
djangorestframework==3.13.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
oauthlib==3.1.1
Pillow==8.4.0
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stripe==2.64.0
tweepy==4.4.0
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
urllib3==1.26.7
zope.interface==5.4.0

pyjwt~=1.7.1
djangorestframework-jwt~=1.11.0